requirements.txt
Django>=1.8
Sphinx
sphinx_rtd_theme
django-bootstrap3
django-annoying==0.10.3
numpy==1.13.1
django-datetime-widget